+++ rust/src/transformations/lipschitz_mul/mod.rs
@@ -131,15 +131,4 @@ impl<const P: usize, Q> LipschitzMulFloatMetric for LpDistance<P, Q> {}
 impl<Q> LipschitzMulFloatMetric for AbsoluteDistance<Q> {}
 
 #[cfg(test)]
-mod test {
-    use super::*;
-
-    #[test]
-    fn test_lipschitz_mul() -> Fallible<()> {
-        let extension =
-            make_lipschitz_float_mul::<AtomDomain<f64>, AbsoluteDistance<f64>>(2., (0., 10.))?;
-        assert_eq!(extension.invoke(&1.3)?, 2.6);
-        println!("{:?}", extension.invoke(&1.3));
-        Ok(())
-    }
-}
+mod test;
